+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Upgrade the `git.setup` module :
  - Add new aliases for `git` commands : `plr`, `rh`, `t`
 - Add a new `lf` alias binary
+- Display binary & module versions when printing the help message
 
 ### Changes